+++ src/Volume.java
@@ -3,16 +3,16 @@ public class Volume {
 	private EnvelopeStage envelopeStage;
 
 	private final static int VOLUME_FADE_PERIOD = 50;
-	private final static float MAX_VOLUME = 1F;
-	private float linearVolume = 0F;
-	private float logarithmicVolume = 0F;
-	private float fadeOutInitFrame;
+	private final static double MAX_VOLUME = 1F;
+	private double linearVolume = 0F;
+	private double logarithmicVolume = 0F;
+	private double fadeOutInitFrame;
 
 	public Volume() {
 		envelopeStage = EnvelopeStage.FADE_IN;
 	}
 
-	public float getVolume(long frames) {
+	public double getVolume(long frames) {
 		updateVolume(frames);
 		return logarithmicVolume;
 	}
@@ -27,22 +27,22 @@ public class Volume {
 	}
 
 	private void updateVolume(long frames) {
-		float fadeSeconds = (float) VOLUME_FADE_PERIOD / 1000;
+		double fadeSeconds = (double) VOLUME_FADE_PERIOD / 1000;
 		int fadeFrames = (int) (fadeSeconds * Sound.SAMPLE_RATE);
-		final float FADE_BASE = (float) Math.E;
+		final double FADE_BASE = (double) Math.E;
 
 		if(envelopeStage == EnvelopeStage.FADE_IN) {
-			float phase = ((float) frames) / fadeFrames;
+			double phase = ((double) frames) / fadeFrames;
 			linearVolume = MAX_VOLUME * phase;
-			logarithmicVolume = (float) ((Math.pow(FADE_BASE, linearVolume) -1) / (FADE_BASE - 1));
+			logarithmicVolume = (double) ((Math.pow(FADE_BASE, linearVolume) -1) / (FADE_BASE - 1));
 			if(linearVolume == MAX_VOLUME) {
 				envelopeStage = EnvelopeStage.PLATEAU;
 			}
 		}
 		if(envelopeStage == EnvelopeStage.FADE_OUT) {
-			float phase = ((float) (frames - fadeOutInitFrame)) / fadeFrames;
+			double phase = ((double) (frames - fadeOutInitFrame)) / fadeFrames;
 			linearVolume = 1 - MAX_VOLUME * phase;
-			logarithmicVolume = (float) ((Math.pow(FADE_BASE, linearVolume) -1) / (FADE_BASE - 1));
+			logarithmicVolume = (double) ((Math.pow(FADE_BASE, linearVolume) -1) / (FADE_BASE - 1));
 			if(linearVolume == 0) {
 				envelopeStage = EnvelopeStage.NULL;
 			}
